move structs to models
